Rename loader-feature module into feature
Make naming more general since it can be used for plugins as well.
