+++ lib/features.js
@@ -13,7 +13,7 @@ const packageHelper = require('./package-helper');
 
 /**
  * An object that holds internal configuration about different
- * "loaders" that can be enabled.
+ * "loaders"/"plugins" that can be enabled/used.
  */
 const loaderFeatures = {
     sass: {
