+++ lib/friendly-errors/formatters/missing-loader.js
@@ -10,7 +10,7 @@
 'use strict';
 
 const chalk = require('chalk');
-const loaderFeatures = require('../../loader-features');
+const loaderFeatures = require('../../features');
 const packageHelper = require('../../package-helper');
 
 function formatErrors(errors) {
