+++ lib/loaders/babel.js
@@ -9,7 +9,7 @@
 
 'use strict';
 
-const loaderFeatures = require('../loader-features');
+const loaderFeatures = require('../features');
 
 /**
  * @param {WebpackConfig} webpackConfig
