+++ lib/loaders/less.js
@@ -9,7 +9,7 @@
 
 'use strict';
 
-const loaderFeatures = require('../loader-features');
+const loaderFeatures = require('../features');
 const cssLoader = require('./css');
 
 /**
