+++ lib/loaders/vue.js
@@ -9,7 +9,7 @@
 
 'use strict';
 
-const loaderFeatures = require('../loader-features');
+const loaderFeatures = require('../features');
 const cssLoaderUtil = require('./css');
 const sassLoaderUtil = require('./sass');
 const lessLoaderUtil = require('./less');
